Update and rename voronoi.py to bionoi.py
